+++ packages/backend/src/nest/storage/notifications/notificationTokens.store.ts
@@ -148,6 +148,12 @@ export class NotificationTokensStore extends EncryptedKeyValueIndexedValidatedSt
   }
 
   public async tombstoneUser(userId: string): Promise<string> {
+    const myEntry = await this.getStore().get(userId)
+    if (!myEntry) {
+      logger.info(`No existing notification token entry found for user ${userId}, must skip tombstone`)
+      return ''
+    }
+
     const tombstoneEntry: PushNotificationTokens = { userId, tokens: [] }
 
     try {