+++ packages/desktop/src/renderer/components/Channel/CreateChannel/CreateChannel.stories.tsx
@@ -18,7 +18,8 @@ const args: CreateChannelProps = {
   createChannel: function (name: string): void {
     logger.info('creating channel: ', name)
   },
-  isAdmin: true,
+  canCreateChannel: true,
+  canCreatePrivateChannel: true,
   handleClose: function (): void {},
   clearErrorsDispatch: function (): void {},
 }