+++ packages/desktop/src/renderer/index.tsx
@@ -3,6 +3,7 @@ import { createRoot, Root as ReactDomRoot } from 'react-dom/client'
 import { ipcRenderer } from 'electron'
 import Root, { persistor } from './Root'
 import store from './store'
+import { clearCommunityWithDependencies } from './clearCommunity'
 import updateHandlers from './store/handlers/update'
 import { socketActions } from './sagas/socket/socket.slice'
 import { communities, captcha } from '@quiet/state-manager'
@@ -74,17 +75,18 @@ if (process.env.NODE_ENV !== 'test') {
   renderApp()
 }
 
-// TODO: this is a bit hacky, should be moved into a saga
-export const clearCommunity = async (remount: boolean = true) => {
-  persistor.pause()
-  await persistor.flush()
-  await persistor.purge()
-  store.dispatch(communities.actions.resetApp('payload'))
-  ipcRenderer.send('clear-community')
-  if (remount) {
-    renderApp()
-  }
-  persistor.persist()
+export const clearCommunity = async () => {
+  await clearCommunityWithDependencies({
+    persistor,
+    dispatch: store.dispatch,
+    resetAppAction: communities.actions.resetApp('payload'),
+    requestBackendLeave: () => ipcRenderer.invoke('clear-community'),
+    remountRoot: () => {
+      root.unmount()
+      root = createRoot(container)
+      root.render(<Root />)
+    },
+  })
 }
 
 if (module.hot) {