+++ packages/mobile/src/screens/CreateChannel/CreateChannel.screen.tsx
@@ -23,7 +23,8 @@ export const CreateChannelScreen: FC = () => {
   const user = useSelector(identity.selectors.currentIdentity)
   const community = useSelector(communities.selectors.currentCommunity)
   const channels = useSelector(publicChannels.selectors.publicChannels)
-  const isOwner = useSelector(communities.selectors.isOwner)
+  const canCreateChannel = useSelector(publicChannels.selectors.canCreateChannel)
+  const canCreatePrivateChannel = useSelector(publicChannels.selectors.canCreatePrivateChannel)
 
   const communityErrors = useSelector(errors.selectors.currentCommunityErrors)
   const error = communityErrors[SocketActions.CREATE_CHANNEL]
@@ -120,8 +121,8 @@ export const CreateChannelScreen: FC = () => {
       channelCreationError={error?.message}
       clearComponent={clearComponent}
       handleBackButton={handleBackButton}
-      // TODO: Pass roles to state manager and check for admin role
-      isAdmin={isOwner}
+      canCreateChannel={canCreateChannel}
+      canCreatePrivateChannel={canCreatePrivateChannel}
     />
   )
 }