+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [unreleased]
+
+### Features
+
+* Allow all users to create private channels [#3277](https://github.com/TryQuiet/quiet/issues/3277)
+* Allow channel members to add members to private channels [#3300](https://github.com/TryQuiet/quiet/issues/3300)
+
+### Fixes
+
+* Don't send deletion message for private channels [#3273](https://github.com/TryQuiet/quiet/issues/3273)
+* Ensure notification registration waits for auth handshake [#3289](https://github.com/TryQuiet/quiet/issues/3289)
+* Fixed a race condition that can cause stale data to remain after leaving community [#3253](https://github.com/TryQuiet/quiet/issues/3253)
+* Temporarily disable private channels
+
+### Chores
+
+* Enable QSS on prod
+
 ## [7.3.0]
 
 ### Features